Avoid duplicates in calls to getSchemaReader
Changed the way members are accumulated in private void axis(...). Now a HashSet is created for every Level instead of an ArrayList. This way duplicates are avoided and a prerequisite of SchemaReader.getMemberChildren is fulfilled (sucha a prerequisite need to be investigated in more detail).
Problems arose when positions were tuples of degree > 1 and thus the same member of a given Hierarchy happened many times in the positions loop. Given a crossjoin of sets with high carcinality created heavy redundancy in members argument of getMemberChildren and it was propagated to the full execution, STRONGLY impacting performance (e.g. case with 2 Hierarchies with 1'000 members each - 1'000'000 of elements had to be scanned insetad of 1'000).